requirements.txt
opencv-python-headless
streamlit
numpy
streamlit-webrtc
tensorflow-cpu
